Split CLI subcommands into dedicated modules
src/cli/lib.rs has grown to thousands of lines and is difficult to
navigate for both humans and agentic tooling.
Split the individual subcommand implementations as well as their
clap `Command` builders into dedicated modules instead.
This should also facilitate refactoring should we choose to switch
from the clap builder API to the derive API.
