+++ src/cli/commands/mod.rs
@@ -1,0 +1,14 @@
+pub mod add_fixmes;
+pub mod analyze;
+pub mod codegen;
+pub mod cyclomatic_complexity;
+pub mod find_executable;
+pub mod find_paths;
+pub mod fix;
+pub mod lint;
+pub mod migrate;
+pub mod migration_candidates;
+pub mod remove_unused_fixmes;
+pub mod security_check;
+pub mod server;
+pub mod test;
