+++ src/cli/lib.rs
@@ -1,42 +1,28 @@
-use clap::{Command, arg};
-use hakana_analyzer::config::{self};
+use clap::Command;
 use hakana_analyzer::custom_hook::CustomHook;
-use hakana_code_info::analysis_result::{
-    AnalysisResult, CheckPointEntry, CheckPointEntryLevel, FullEntry, HhClientEntry,
-};
-use hakana_code_info::data_flow::graph::{GraphKind, WholeProgramKind};
-use hakana_code_info::issue::IssueKind;
-use hakana_language_server::server_client::ServerConnection;
-use hakana_protocol::ClientSocket;
-use hakana_str::Interner;
-use indexmap::IndexMap;
-use indicatif::{ProgressBar, ProgressStyle};
-use line_break_map::LineBreakMap;
 use log::LevelFilter;
-use rand::Rng;
 use rustc_hash::FxHashSet;
-use serde_json::json;
-use std::collections::BTreeMap;
 use std::env;
-use std::fs::{self, File};
-use std::io::{IsTerminal, Write};
-use std::path::{Path, PathBuf};
+use std::fs;
+use std::io::IsTerminal;
+use std::path::Path;
 use std::process::exit;
-use std::sync::Arc;
 use test_runners::test_runner::TestRunner;
-
-pub mod mcp;
-pub mod test_runners;
 
 /// Print a message only when stdout is a terminal (not piped).
 /// Use this for informational/summary output that should not appear in piped output.
 macro_rules! tty_println {
     ($($arg:tt)*) => {
-        if std::io::stdout().is_terminal() {
+        if <std::io::Stdout as std::io::IsTerminal>::is_terminal(&std::io::stdout()) {
             println!($($arg)*);
         }
     };
 }
+
+pub mod commands;
+pub mod helpers;
+pub mod mcp;
+pub mod test_runners;
 
 pub async fn init(
     analysis_hooks: Vec<Box<dyn CustomHook>>,
@@ -61,552 +47,25 @@ pub async fn init(
         .map(|i| i.to_string())
         .collect::<FxHashSet<_>>();
 
-    let matches =
-        Command::new("hakana")
-            .about("Another static analysis tool for Hack")
-            .subcommand_required(true)
-            .arg_required_else_help(true)
-            .subcommand(
-                Command::new("analyze")
-                    .alias("analyse")
-                    .about("Analyzes code in the current directory")
-                    .arg(arg!(--"root" <PATH>).required(false).help(
-                        "The root directory that Hakana runs in. Defaults to the current directory",
-                    ))
-                    .arg(
-                        arg!(--"config" <PATH>)
-                            .required(false)
-                            .help("Hakana config path — defaults to ./hakana.json"),
-                    )
-                    .arg(
-                        arg!(--"filter" <PATH>)
-                            .required(false)
-                            .help("Filter the files that are analyzed"),
-                    )
-                    .arg(
-                        arg!(--"ignore" <PATH>)
-                            .required(false)
-                            .multiple(true)
-                            .help("Ignore certain files during analysis"),
-                    )
-                    .arg(
-                        arg!(--"threads" <PATH>)
-                            .required(false)
-                            .help("How many threads to use"),
-                    )
-                    .arg(
-                        arg!(--"find-unused-expressions")
-                            .required(false)
-                            .help("Find unused expressions"),
-                    )
-                    .arg(
-                        arg!(--"find-unused-definitions")
-                            .required(false)
-                            .help("Find unused definitions — classes, functions, methods etc."),
-                    )
-                    .arg(
-                        arg!(--"show-issue" <PATH>)
-                            .required(false)
-                            .multiple(true)
-                            .help("Only output issues of this/these type(s)"),
-                    )
-                    .arg(arg!(--"all-issues").required(false).help("Show all issues"))
-                    .arg(
-                        arg!(--"ignore-mixed-issues")
-                            .required(false)
-                            .help("Ignore mixed/any issues"),
-                    )
-                    .arg(
-                        arg!(--"show-mixed-function-counts")
-                            .required(false)
-                            .help("Show which functions we lead to mixed types"),
-                    )
-                    .arg(
-                        arg!(--"print-symbol-usages")
-                            .required(false)
-                            .help("Output a JSON map of symbol definitions and usages"),
-                    )
-                    .arg(
-                        arg!(--"debug")
-                            .required(false)
-                            .help("Add output for debugging"),
-                    )
-                    .arg(
-                        arg!(--"show-timing")
-                            .required(false)
-                            .help("If set, timing info will be displayed"),
-                    )
-                    .arg(
-                        arg!(--"no-cache")
-                            .required(false)
-                            .help("Whether to ignore the cache"),
-                    )
-                    .arg(
-                        arg!(--"diff")
-                            .required(false)
-                            .help("Whether perform AST-based diffing to speed up execution"),
-                    )
-                    .arg(
-                        arg!(--"show-issue-stats")
-                            .required(false)
-                            .help("Output a summary of issue counts"),
-                    )
-                    .arg(
-                        arg!(--"output" <PATH>)
-                            .required(false)
-                            .help("File to save output to"),
-                    )
-                    .arg(arg!(--"json-format" <FORMAT>).required(false).help(
-                        "Format for JSON output. Options: checkpoint (default), full, hh_client",
-                    ))
-                    .arg(
-                        arg!(--"standalone")
-                            .required(false)
-                            .help("Run analysis directly without connecting to server (default for CI)"),
-                    )
-                    .arg(
-                        arg!(--"with-server")
-                            .required(false)
-                            .help("Use server mode: connect to existing server or spawn one if needed"),
-                    ),
-            )
-            .subcommand(
-                Command::new("migration-candidates")
-                    .about("Generates a list of all migration candidates")
-                    .arg(arg!(--"root" <PATH>).required(false).help(
-                        "The root directory that Hakana runs in. Defaults to the current directory",
-                    ))
-                    .arg(
-                        arg!(--"config" <PATH>)
-                            .required(false)
-                            .help("Hakana config path — defaults to ./hakana.json"),
-                    )
-                    .arg(
-                        arg!(--"migration" <PATH>)
-                            .required(true)
-                            .help("The migration you want to perform"),
-                    )
-                    .arg(
-                        arg!(--"threads" <PATH>)
-                            .required(false)
-                            .help("How many threads to use"),
-                    )
-                    .arg(
-                        arg!(--"filter" <PATH>)
-                            .required(false)
-                            .help("Filter the files that are analyzed"),
-                    )
-                    .arg(
-                        arg!(--"debug")
-                            .required(false)
-                            .help("Add output for debugging"),
-                    ),
-            )
-            .subcommand(
-                Command::new("codegen")
-                    .about("Generates codegen")
-                    .arg(arg!(--"root" <PATH>).required(false).help(
-                        "The root directory that Hakana runs in. Defaults to the current directory",
-                    ))
-                    .arg(
-                        arg!(--"config" <PATH>)
-                            .required(false)
-                            .help("Hakana config path — defaults to ./hakana.json"),
-                    )
-                    .arg(arg!(--"name" <PATH>).required(false).help(
-                        "The codegen you want to perform — if omitted, all codegen is generated",
-                    ))
-                    .arg(
-                        arg!(--"check")
-                            .required(false)
-                            .help("If passed, will just verify that codegen is accurate"),
-                    )
-                    .arg(
-                        arg!(--"overwrite")
-                            .required(false)
-                            .help("If passed, will overwrite any conflicting files"),
-                    )
-                    .arg(
-                        arg!(--"threads" <PATH>)
-                            .required(false)
-                            .help("How many threads to use"),
-                    )
-                    .arg(
-                        arg!(--"debug")
-                            .required(false)
-                            .help("Add output for debugging"),
-                    )
-                    .arg(
-                        arg!(--"output" <PATH>)
-                            .required(false)
-                            .help("File to save output to"),
-                    ),
-            )
-            .subcommand(
-                Command::new("migrate")
-                    .about("Migrates code in the current directory")
-                    .arg(arg!(--"root" <PATH>).required(false).help(
-                        "The root directory that Hakana runs in. Defaults to the current directory",
-                    ))
-                    .arg(
-                        arg!(--"config" <PATH>)
-                            .required(false)
-                            .help("Hakana config path — defaults to ./hakana.json"),
-                    )
-                    .arg(
-                        arg!(--"migration" <PATH>)
-                            .required(true)
-                            .help("The migration you want to perform"),
-                    )
-                    .arg(
-                        arg!(--"symbols" <PATH>)
-                            .required(true)
-                            .help("The path to a list of symbols, separated by newlines"),
-                    )
-                    .arg(
-                        arg!(--"filter" <PATH>)
-                            .required(false)
-                            .help("Filter the files that are analyzed"),
-                    )
-                    .arg(
-                        arg!(--"threads" <PATH>)
-                            .required(false)
-                            .help("How many threads to use"),
-                    )
-                    .arg(
-                        arg!(--"debug")
-                            .required(false)
-                            .help("Add output for debugging"),
-                    ),
-            )
-            .subcommand(
-                Command::new("add-fixmes")
-                    .about("Adds fixmes to suppress Hakana issues")
-                    .arg(arg!(--"root" <PATH>).required(false).help(
-                        "The root directory that Hakana runs in. Defaults to the current directory",
-                    ))
-                    .arg(
-                        arg!(--"config" <PATH>)
-                            .required(false)
-                            .help("Hakana config path — defaults to ./hakana.json"),
-                    )
-                    .arg(
-                        arg!(--"issue" <PATH>)
-                            .required(true)
-                            .multiple(true)
-                            .help("The issue or issues to add fixmes for"),
-                    )
-                    .arg(
-                        arg!(--"filter" <PATH>)
-                            .required(false)
-                            .help("Filter the files that have added fixmes"),
-                    )
-                    .arg(
-                        arg!(--"threads" <PATH>)
-                            .required(false)
-                            .help("How many threads to use"),
-                    )
-                    .arg(
-                        arg!(--"debug")
-                            .required(false)
-                            .help("Add output for debugging"),
-                    ),
-            )
-            .subcommand(
-                Command::new("remove-unused-fixmes")
-                    .about("Removes all fixmes that are never used")
-                    .arg(arg!(--"root" <PATH>).required(false).help(
-                        "The root directory that Hakana runs in. Defaults to the current directory",
-                    ))
-                    .arg(
-                        arg!(--"config" <PATH>)
-                            .required(false)
-                            .help("Hakana config path — defaults to ./hakana.json"),
-                    )
-                    .arg(
-                        arg!(--"threads" <PATH>)
-                            .required(false)
-                            .help("How many threads to use"),
-                    )
-                    .arg(
-                        arg!(--"filter" <PATH>)
-                            .required(false)
-                            .help("Filter the files that have added fixmes"),
-                    )
-                    .arg(
-                        arg!(--"debug")
-                            .required(false)
-                            .help("Add output for debugging"),
-                    ),
-            )
-            .subcommand(
-                Command::new("fix")
-                    .about("Fixes issues in the codebase")
-                    .arg(arg!(--"root" <PATH>).required(false).help(
-                        "The root directory that Hakana runs in. Defaults to the current directory",
-                    ))
-                    .arg(
-                        arg!(--"config" <PATH>)
-                            .required(false)
-                            .help("Hakana config path — defaults to ./hakana.json"),
-                    )
-                    .arg(
-                        arg!(--"issue" <PATH>)
-                            .required(true)
-                            .help("The issue to fix"),
-                    )
-                    .arg(
-                        arg!(--"filter" <PATH>)
-                            .required(false)
-                            .help("Filter the files that are fixed"),
-                    )
-                    .arg(
-                        arg!(--"threads" <PATH>)
-                            .required(false)
-                            .help("How many threads to use"),
-                    )
-                    .arg(
-                        arg!(--"debug")
-                            .required(false)
-                            .help("Add output for debugging"),
-                    ),
-            )
-            .subcommand(
-                Command::new("security-check")
-                    .about("Looks for vulnerabilities in the codebase")
-                    .arg(arg!(--"root" <PATH>).required(false).help(
-                        "The root directory that Hakana runs in. Defaults to the current directory",
-                    ))
-                    .arg(
-                        arg!(--"config" <PATH>)
-                            .required(false)
-                            .help("Hakana config path — defaults to ./hakana.json"),
-                    )
-                    .arg(
-                        arg!(--"threads" <PATH>)
-                            .required(false)
-                            .help("How many threads to use"),
-                    )
-                    .arg(
-                        arg!(--"max-depth" <PATH>)
-                            .required(false)
-                            .help("Length of the longest allowable path — defaults to 20, and overrides config file value"),
-                    )
-                    .arg(
-                        arg!(--"debug")
-                            .required(false)
-                            .help("Add output for debugging"),
-                    )
-                    .arg(
-                        arg!(--"output" <PATH>)
-                            .required(false)
-                            .help("File to save output to"),
-                    ),
-            )
-            .subcommand(
-                Command::new("find-paths")
-                    .about("Does whole-program analysis querying")
-                    .arg(arg!(--"root" <PATH>).required(false).help(
-                        "The root directory that Hakana runs in. Defaults to the current directory",
-                    ))
-                    .arg(
-                        arg!(--"config" <PATH>)
-                            .required(false)
-                            .help("Hakana config path — defaults to ./hakana.json"),
-                    )
-                    .arg(
-                        arg!(--"threads" <PATH>)
-                            .required(false)
-                            .help("How many threads to use"),
-                    )
-                    .arg(
-                        arg!(--"max-depth" <PATH>)
-                            .required(false)
-                            .help("Length of the longest allowable path — defaults to 20, and overrides config file value"),
-                    )
-                    .arg(
-                        arg!(--"debug")
-                            .required(false)
-                            .help("Add output for debugging"),
-                    ),
-            )
-            .subcommand(
-                Command::new("test")
-                    .about("Runs one or more Hakana tests")
-                    .arg(
-                        arg!(--"no-cache")
-                            .required(false)
-                            .help("Whether to use cache"),
-                    )
-                    .arg(
-                        arg!(--"reuse-codebase")
-                            .required(false)
-                            .help("Whether to reuse codebase between tests"),
-                    )
-                    .arg(
-                        arg!(--"randomize")
-                            .required(false)
-                            .help("Whether to randomise test order"),
-                    )
-                    .arg(
-                        arg!(--"seed" <COUNT>)
-                            .required(false)
-                            .help("Seed for random test execution"),
-                    )
-                    .arg(
-                        arg!(--"debug")
-                            .required(false)
-                            .help("Whether to show debug output"),
-                    )
-                    .arg(
-                        arg!(--"repeat" <COUNT>)
-                            .required(false)
-                            .help("How many times to repeat the test (useful for profiling)"),
-                    )
-                    .arg(arg!(<TEST> "The test to run"))
-                    .arg_required_else_help(true),
-            )
-            .subcommand(
-                Command::new("find-executable")
-                    .about("Finds all executable lines of code")
-                    .arg(arg!(--"root" <PATH>).required(false).help(
-                        "The root directory that Hakana runs in. Defaults to the current directory",
-                    ))
-                    .arg(
-                        arg!(--"output" <PATH>)
-                            .required(true)
-                            .help("File to save output to"),
-                    ),
-            )
-            .subcommand(
-                Command::new("lint")
-                    .about("Runs linters on Hack code (HHAST-compatible)")
-                    .arg(arg!(--"root" <PATH>).required(false).help(
-                        "The root directory that Hakana runs in. Defaults to the current directory",
-                    ))
-                    .arg(
-                        arg!(--"config" <PATH>)
-                            .required(false)
-                            .help("Path to hhast-lint.json config file — defaults to ./hhast-lint.json"),
-                    )
-                    .arg(
-                        arg!(--"threads" <COUNT>)
-                            .required(false)
-                            .help("How many threads to use"),
-                    )
-                    .arg(
-                        arg!(--"fix")
-                            .required(false)
-                            .help("Apply auto-fixes where available"),
-                    )
-                    .arg(
-                        arg!(--"add-fixmes")
-                            .required(false)
-                            .help("Add HHAST_FIXME comments for linter issues"),
-                    )
-                    .arg(
-                        arg!(--"linter" <NAME>)
-                            .required(false)
-                            .multiple(true)
-                            .help("Run specific linter(s) by HHAST name or kebab-case name. When used with --add-fixmes, only add fixmes for these linter(s)"),
-                    )
-                    .arg(
-                        arg!(--"no-codeowners")
-                            .required(false)
-                            .help("Skip files that have codeowners (listed in CODEOWNERS file)"),
-                    )
-                    .arg(
-                        arg!(--"debug-timings")
-                            .required(false)
-                            .help("Show timing information for each linter"),
-                    )
-                    .arg(
-                        arg!(--"debug")
-                            .required(false)
-                            .help("Add output for debugging"),
-                    )
-                    .arg(
-                        arg!(--"output" <PATH>)
-                            .required(false)
-                            .help("File to save output to"),
-                    )
-                    .arg(
-                        arg!([PATH] "Optional file or directory to lint (defaults to config roots)")
-                            .required(false)
-                            .multiple(true),
-                    ),
-            )
-            .subcommand(
-                Command::new("cyclomatic-complexity")
-                    .about("Analyzes cyclomatic complexity of functions in the codebase")
-                    .arg(arg!(--"root" <PATH>).required(false).help(
-                        "The root directory that Hakana runs in. Defaults to the current directory",
-                    ))
-                    .arg(
-                        arg!(--"config" <PATH>)
-                            .required(false)
-                            .help("Hakana config path — defaults to ./hakana.json"),
-                    )
-                    .arg(
-                        arg!(--"filter" <PATH>)
-                            .required(false)
-                            .multiple(true)
-                            .help("Filter the files that are analyzed (glob patterns)"),
-                    )
-                    .arg(
-                        arg!(--"threshold" <NUM>)
-                            .required(false)
-                            .help("Complexity threshold to report (default: 10)"),
-                    )
-                    .arg(
-                        arg!(--"threads" <PATH>)
-                            .required(false)
-                            .help("How many threads to use"),
-                    )
-                    .arg(
-                        arg!(--"output" <PATH>)
-                            .required(false)
-                            .help("File to save JSON output to"),
-                    )
-                    .arg(
-                        arg!(--"debug")
-                            .required(false)
-                            .help("Add output for debugging"),
-                    ),
-            )
-            .subcommand(
-                Command::new("server")
-                    .about("Start or manage the hakana server")
-                    .arg(arg!(--"root" <PATH>).required(false).help(
-                        "The root directory that Hakana runs in. Defaults to the current directory",
-                    ))
-                    .arg(
-                        arg!(--"config" <PATH>)
-                            .required(false)
-                            .help("Hakana config path — defaults to ./hakana.json"),
-                    )
-                    .arg(
-                        arg!(--"threads" <PATH>)
-                            .required(false)
-                            .help("How many threads to use"),
-                    )
-                    .arg(
-                        arg!(--"debug")
-                            .required(false)
-                            .help("Add output for debugging"),
-                    )
-                    .arg(
-                        arg!(--"stop")
-                            .required(false)
-                            .help("Stop a running server"),
-                    )
-                    .arg(
-                        arg!(--"status")
-                            .required(false)
-                            .help("Show server status"),
-                    ),
-            )
-            .get_matches();
+    let matches = Command::new("hakana")
+        .about("Another static analysis tool for Hack")
+        .subcommand_required(true)
+        .arg_required_else_help(true)
+        .subcommand(commands::analyze::get_subcommand())
+        .subcommand(commands::migration_candidates::get_subcommand())
+        .subcommand(commands::codegen::get_subcommand())
+        .subcommand(commands::migrate::get_subcommand())
+        .subcommand(commands::add_fixmes::get_subcommand())
+        .subcommand(commands::remove_unused_fixmes::get_subcommand())
+        .subcommand(commands::fix::get_subcommand())
+        .subcommand(commands::security_check::get_subcommand())
+        .subcommand(commands::find_paths::get_subcommand())
+        .subcommand(commands::test::get_subcommand())
+        .subcommand(commands::find_executable::get_subcommand())
+        .subcommand(commands::lint::get_subcommand())
+        .subcommand(commands::cyclomatic_complexity::get_subcommand())
+        .subcommand(commands::server::get_subcommand())
+        .get_matches();
 
     let cwd = (env::current_dir()).unwrap().to_str().unwrap().to_string();
 
@@ -704,7 +163,7 @@ pub async fn init(
 
     match matches.subcommand() {
         Some(("analyze", sub_matches)) => {
-            do_analysis(
+            commands::analyze::handle(
                 sub_matches,
                 all_custom_issues,
                 &root_dir,
@@ -720,7 +179,7 @@ pub async fn init(
             .await;
         }
         Some(("security-check", sub_matches)) => {
-            do_security_check(
+            commands::security_check::handle(
                 &cwd,
                 all_custom_issues,
                 config_path,
@@ -733,7 +192,7 @@ pub async fn init(
             );
         }
         Some(("find-paths", sub_matches)) => {
-            do_find_paths(
+            commands::find_paths::handle(
                 &cwd,
                 all_custom_issues,
                 config_path,
@@ -746,7 +205,7 @@ pub async fn init(
             );
         }
         Some(("migrate", sub_matches)) => {
-            do_migrate(
+            commands::migrate::handle(
                 sub_matches,
                 &root_dir,
                 all_custom_issues,
@@ -759,7 +218,7 @@ pub async fn init(
             );
         }
         Some(("migration-candidates", sub_matches)) => {
-            do_migration_candidates(
+            commands::migration_candidates::handle(
                 sub_matches,
                 &root_dir,
                 all_custom_issues,
@@ -772,7 +231,7 @@ pub async fn init(
             );
         }
         Some(("codegen", sub_matches)) => {
-            do_codegen(
+            commands::codegen::handle(
                 sub_matches,
                 &root_dir,
                 all_custom_issues,
@@ -786,7 +245,7 @@ pub async fn init(
             );
         }
         Some(("add-fixmes", sub_matches)) => {
-            do_add_fixmes(
+            commands::add_fixmes::handle(
                 sub_matches,
                 all_custom_issues,
                 &root_dir,
@@ -799,7 +258,7 @@ pub async fn init(
             );
         }
         Some(("remove-unused-fixmes", sub_matches)) => {
-            do_remove_unused_fixmes(
+            commands::remove_unused_fixmes::handle(
                 sub_matches,
                 &root_dir,
                 all_custom_issues,
@@ -812,7 +271,7 @@ pub async fn init(
             );
         }
         Some(("fix", sub_matches)) => {
-            do_fix(
+            commands::fix::handle(
                 sub_matches,
                 all_custom_issues,
                 analysis_hooks,
@@ -825,45 +284,25 @@ pub async fn init(
             );
         }
         Some(("test", sub_matches)) => {
-            let repeat = if let Some(val) = sub_matches.value_of("repeat").map(|f| f.to_string()) {
-                val.parse::<u16>().unwrap()
-            } else {
-                0
-            };
-
-            let random_seed = if sub_matches.is_present("randomize") {
-                if let Some(val) = sub_matches.value_of("seed").map(|f| f.to_string()) {
-                    Some(val.parse::<u64>().unwrap())
-                } else {
-                    let mut rng = rand::thread_rng();
-                    Some(rng.r#gen())
-                }
-            } else {
-                None
-            };
-
-            test_runner.run_test(
-                sub_matches.value_of("TEST").expect("required").to_string(),
+            commands::test::handle(
+                sub_matches,
+                test_runner,
                 show_progress,
-                !sub_matches.is_present("no-cache"),
-                sub_matches.is_present("reuse-codebase"),
                 &mut had_error,
                 header,
-                repeat,
-                random_seed,
             );
         }
         Some(("find-executable", sub_matches)) => {
-            do_find_executable(sub_matches, &root_dir, &cwd, threads, show_progress);
+            commands::find_executable::handle(sub_matches, &root_dir, &cwd, threads, show_progress);
         }
         Some(("lint", sub_matches)) => {
-            do_lint(sub_matches, &root_dir, &mut had_error, custom_linters);
+            commands::lint::handle(sub_matches, &root_dir, &mut had_error, custom_linters);
         }
         Some(("server", sub_matches)) => {
-            do_server(sub_matches, &root_dir, threads, header, analysis_hooks).await;
+            commands::server::handle(sub_matches, &root_dir, threads, header, analysis_hooks).await;
         }
         Some(("cyclomatic-complexity", sub_matches)) => {
-            do_cyclomatic_complexity(
+            commands::cyclomatic_complexity::handle(
                 sub_matches,
                 all_custom_issues,
                 &root_dir,
@@ -875,2083 +314,10 @@ pub async fn init(
                 &mut had_error,
             );
         }
-        _ => unreachable!(), // If all subcommands are defined above, anything else is unreachable!()
+        _ => unreachable!(),
     }
 
     if had_error {
         exit(1);
-    }
-}
-
-fn do_fix(
-    sub_matches: &clap::ArgMatches,
-    all_custom_issues: FxHashSet<String>,
-    analysis_hooks: Vec<Box<dyn CustomHook>>,
-    root_dir: String,
-    config_path: Option<&Path>,
-    cwd: String,
-    threads: u8,
-    show_progress: bool,
-    header: &str,
-) {
-    let issue_name = sub_matches.value_of("issue").unwrap().to_string();
-    let issue_kind = IssueKind::from_str_custom(&issue_name, &all_custom_issues).unwrap();
-
-    if !matches!(issue_kind, IssueKind::CustomIssue(..)) && !issue_kind.has_autofix() {
-        println!("Issue type {} does not support autofixing", issue_kind);
-        exit(1);
-    }
-
-    let filter = sub_matches.value_of("filter").map(|f| f.to_string());
-
-    let mut config = config::Config::new(root_dir.clone(), all_custom_issues);
-    config.hooks = analysis_hooks.into_iter().map(Arc::from).collect();
-
-    config.find_unused_expressions = issue_kind.requires_dataflow_analysis();
-    config.find_unused_definitions = issue_kind.is_unused_definition();
-    config.issues_to_fix.insert(issue_kind);
-
-    let config_path = config_path.unwrap();
-
-    let mut interner = Interner::default();
-
-    if config_path.exists() {
-        config
-            .update_from_file(&cwd, config_path, &mut interner)
-            .ok();
-    }
-
-    config.allowed_issues = None;
-
-    let result = hakana_orchestrator::scan_and_analyze(
-        Vec::new(),
-        filter,
-        None,
-        Arc::new(config),
-        None,
-        threads,
-        show_progress,
-        header,
-        Arc::new(interner),
-        None,
-        None,
-        None,
-        || {},
-    );
-
-    if let Ok((mut analysis_result, successfull_run_data)) = result {
-        update_files(
-            &mut analysis_result,
-            &root_dir,
-            &successfull_run_data.interner,
-        );
-    }
-}
-
-fn do_find_executable(
-    sub_matches: &clap::ArgMatches,
-    root_dir: &str,
-    cwd: &String,
-    threads: u8,
-    show_progress: bool,
-) {
-    let output_file = sub_matches.value_of("output").unwrap().to_string();
-    let config = config::Config::new(root_dir.to_string(), FxHashSet::default());
-
-    match executable_finder::scan_files(
-        &vec![root_dir.to_string()],
-        None,
-        &Arc::new(config),
-        threads,
-        show_progress,
-    ) {
-        Ok(file_infos) => {
-            let output_path = if output_file.starts_with('/') {
-                output_file
-            } else {
-                format!("{}/{}", cwd, output_file)
-            };
-            let mut out = fs::File::create(Path::new(&output_path)).unwrap();
-            match write!(
-                out,
-                "{}",
-                serde_json::to_string_pretty(&file_infos).unwrap()
-            ) {
-                Ok(_) => {
-                    tty_println!("Done")
-                }
-                Err(err) => {
-                    println!("error: {}", err)
-                }
-            }
-        }
-        Err(_) => {
-            println!("error")
-        }
-    }
-}
-
-fn do_remove_unused_fixmes(
-    sub_matches: &clap::ArgMatches,
-    root_dir: &String,
-    all_custom_issues: FxHashSet<String>,
-    analysis_hooks: Vec<Box<dyn CustomHook>>,
-    config_path: Option<&Path>,
-    cwd: &String,
-    threads: u8,
-    show_progress: bool,
-    header: &str,
-) {
-    let filter = sub_matches.value_of("filter").map(|f| f.to_string());
-
-    let mut config = config::Config::new(root_dir.clone(), all_custom_issues);
-
-    config.hooks = analysis_hooks.into_iter().map(Arc::from).collect();
-
-    let config_path = config_path.unwrap();
-
-    let mut interner = Interner::default();
-
-    if config_path.exists() {
-        config
-            .update_from_file(cwd, config_path, &mut interner)
-            .ok();
-    }
-    config.allowed_issues = None;
-
-    config.find_unused_expressions = true;
-    config.find_unused_definitions = true;
-
-    config.remove_fixmes = true;
-
-    let result = hakana_orchestrator::scan_and_analyze(
-        Vec::new(),
-        filter,
-        None,
-        Arc::new(config),
-        None,
-        threads,
-        show_progress,
-        header,
-        Arc::new(interner),
-        None,
-        None,
-        None,
-        || {},
-    );
-
-    if let Ok((mut analysis_result, successful_run_data)) = result {
-        update_files(
-            &mut analysis_result,
-            root_dir,
-            &successful_run_data.interner,
-        );
-    }
-}
-
-fn do_add_fixmes(
-    sub_matches: &clap::ArgMatches,
-    all_custom_issues: FxHashSet<String>,
-    root_dir: &String,
-    analysis_hooks: Vec<Box<dyn CustomHook>>,
-    config_path: Option<&Path>,
-    cwd: &String,
-    threads: u8,
-    show_progress: bool,
-    header: &str,
-) {
-    let filter_issue_strings = sub_matches
-        .values_of("issue")
-        .map(|values| values.collect::<FxHashSet<_>>());
-
-    let mut issue_kinds_filter = FxHashSet::default();
-
-    if let Some(filter_issue_strings) = filter_issue_strings {
-        for filter_issue_string in filter_issue_strings {
-            if let Ok(issue_kind) =
-                IssueKind::from_str_custom(filter_issue_string, &all_custom_issues)
-            {
-                issue_kinds_filter.insert(issue_kind);
-            } else {
-                println!("Invalid issue type {}", filter_issue_string);
-                exit(1);
-            }
-        }
-    }
-
-    let filter = sub_matches.value_of("filter").map(|f| f.to_string());
-
-    let mut config = config::Config::new(root_dir.clone(), all_custom_issues);
-
-    config.issues_to_fix.extend(issue_kinds_filter);
-    config.hooks = analysis_hooks.into_iter().map(Arc::from).collect();
-    config.find_unused_expressions = true;
-    config.find_unused_definitions = true;
-
-    let config_path = config_path.unwrap();
-
-    let mut interner = Interner::default();
-
-    if config_path.exists() {
-        config
-            .update_from_file(cwd, config_path, &mut interner)
-            .ok();
-    }
-    config.allowed_issues = None;
-
-    config.add_fixmes = true;
-
-    let result = hakana_orchestrator::scan_and_analyze(
-        Vec::new(),
-        filter,
-        None,
-        Arc::new(config),
-        None,
-        threads,
-        show_progress,
-        header,
-        Arc::new(interner),
-        None,
-        None,
-        None,
-        || {},
-    );
-
-    if let Ok((mut analysis_result, successful_run_data)) = result {
-        update_files(
-            &mut analysis_result,
-            root_dir,
-            &successful_run_data.interner,
-        );
-    }
-}
-
-fn do_migrate(
-    sub_matches: &clap::ArgMatches,
-    root_dir: &String,
-    all_custom_issues: FxHashSet<String>,
-    migration_hooks: Vec<Box<dyn CustomHook>>,
-    config_path: Option<&Path>,
-    cwd: &String,
-    threads: u8,
-    show_progress: bool,
-    header: &str,
-) {
-    let migration_name = sub_matches.value_of("migration").unwrap().to_string();
-    let migration_source = sub_matches.value_of("symbols").unwrap().to_string();
-
-    let mut config = config::Config::new(root_dir.clone(), all_custom_issues);
-    config.hooks = migration_hooks
-        .into_iter()
-        .filter(|m| {
-            if let Some(name) = m.get_migration_name() {
-                migration_name == name
-            } else {
-                false
-            }
-        })
-        .map(Arc::from)
-        .collect();
-
-    if config.hooks.is_empty() {
-        println!("Migration {} not recognised", migration_name);
-        exit(1);
-    }
-
-    let config_path = config_path.unwrap();
-
-    let mut interner = Interner::default();
-
-    if config_path.exists() {
-        config
-            .update_from_file(cwd, config_path, &mut interner)
-            .ok();
-    }
-    config.allowed_issues = None;
-
-    let file_path = format!("{}/{}", cwd, migration_source);
-
-    let buf = fs::read_to_string(file_path.clone());
-
-    if let Ok(contents) = buf {
-        config.migration_symbols = contents
-            .lines()
-            .map(|v| {
-                let mut parts = v.split(',').collect::<Vec<_>>();
-                let first_part = parts.remove(0);
-                (first_part.to_string(), parts.join(","))
-            })
-            .collect();
-        config.in_migration = true;
-    } else {
-        println!(
-            "\nERROR: File {} does not exist or could not be read\n",
-            file_path
-        );
-        exit(1);
-    }
-
-    let filter = sub_matches.value_of("filter").map(|f| f.to_string());
-
-    let result = hakana_orchestrator::scan_and_analyze(
-        Vec::new(),
-        filter,
-        None,
-        Arc::new(config),
-        None,
-        threads,
-        show_progress,
-        header,
-        Arc::new(interner),
-        None,
-        None,
-        None,
-        || {},
-    );
-
-    if let Ok((mut analysis_result, successful_run_data)) = result {
-        update_files(
-            &mut analysis_result,
-            root_dir,
-            &successful_run_data.interner,
-        );
-    }
-}
-
-fn do_migration_candidates(
-    sub_matches: &clap::ArgMatches,
-    root_dir: &str,
-    all_custom_issues: FxHashSet<String>,
-    migration_hooks: Vec<Box<dyn CustomHook>>,
-    config_path: Option<&Path>,
-    cwd: &String,
-    threads: u8,
-    show_progress: bool,
-    header: &str,
-) {
-    let migration_name = sub_matches.value_of("migration").unwrap().to_string();
-
-    let mut config = config::Config::new(root_dir.to_string(), all_custom_issues);
-    config.hooks = migration_hooks
-        .into_iter()
-        .filter(|m| {
-            if let Some(name) = m.get_migration_name() {
-                migration_name == name
-            } else {
-                false
-            }
-        })
-        .map(Arc::from)
-        .collect();
-
-    config.in_migration = true;
-
-    if config.hooks.is_empty() {
-        println!("Migration {} not recognised", migration_name);
-        exit(1);
-    }
-
-    let config_path = config_path.unwrap();
-
-    let mut interner = Interner::default();
-
-    if config_path.exists() {
-        config
-            .update_from_file(cwd, config_path, &mut interner)
-            .ok();
-    }
-    config.allowed_issues = None;
-
-    let config = Arc::new(config);
-
-    let filter = sub_matches.value_of("filter").map(|f| f.to_string());
-
-    let result = hakana_orchestrator::scan_and_analyze(
-        Vec::new(),
-        filter,
-        None,
-        config.clone(),
-        None,
-        threads,
-        show_progress,
-        header,
-        Arc::new(interner),
-        None,
-        None,
-        None,
-        || {},
-    );
-
-    if let Ok(result) = result {
-        tty_println!("\nSymbols to migrate:\n");
-        for config_hook in &config.hooks {
-            let migration_candidates =
-                config_hook.get_candidates(&result.1.codebase, &result.1.interner, &result.0);
-
-            for migration_candidate in migration_candidates {
-                println!("{}", migration_candidate);
-            }
-        }
-    }
-}
-
-fn do_codegen(
-    sub_matches: &clap::ArgMatches,
-    root_dir: &str,
-    all_custom_issues: FxHashSet<String>,
-    analysis_hooks: Vec<Box<dyn CustomHook>>,
-    mut codegen_hooks: Vec<Box<dyn CustomHook>>,
-    config_path: Option<&Path>,
-    cwd: &String,
-    threads: u8,
-    show_progress: bool,
-    header: &str,
-) {
-    let codegen_name = sub_matches.value_of("name");
-    let check_codegen = sub_matches.is_present("check");
-    let overwrite_codegen = sub_matches.is_present("overwrite");
-
-    let output_file = sub_matches.value_of("output").map(|f| f.to_string());
-
-    let mut config = config::Config::new(root_dir.to_string(), all_custom_issues);
-    config.hooks = analysis_hooks.into_iter().map(Arc::from).collect();
-    config.in_codegen = true;
-
-    if let Some(codegen_name) = codegen_name {
-        codegen_hooks.retain(|hook| {
-            if let Some(candidate_name) = hook.get_codegen_name() {
-                candidate_name == codegen_name
-            } else {
-                false
-            }
-        });
-
-        if codegen_hooks.is_empty() {
-            println!("Codegen {} not recognised", codegen_name);
-            exit(1);
-        }
-    }
-
-    config
-        .hooks
-        .extend(codegen_hooks.into_iter().map(Arc::from));
-
-    let config_path = config_path.unwrap();
-
-    let mut interner = Interner::default();
-
-    if config_path.exists() {
-        config
-            .update_from_file(cwd, config_path, &mut interner)
-            .ok();
-    }
-    config.allowed_issues = None;
-
-    let config = Arc::new(config);
-
-    let result = hakana_orchestrator::scan_and_analyze(
-        Vec::new(),
-        None,
-        None,
-        config.clone(),
-        None,
-        threads,
-        show_progress,
-        header,
-        Arc::new(interner),
-        None,
-        None,
-        None,
-        || {},
-    );
-
-    if let Ok(result) = result {
-        let mut errors = vec![];
-        let mut updated_count = 0;
-        let mut verified_count = 0;
-
-        let mut already_written_files = FxHashSet::default();
-
-        for (name, info) in result.0.codegen {
-            if already_written_files.contains(&name) {
-                errors.push((name.clone(), "File already written".to_string()));
-                continue;
-            }
-
-            already_written_files.insert(name.clone());
-
-            let path = Path::new(&name);
-            if !path.exists() {
-                if check_codegen {
-                    errors.push((name.clone(), "Doesn’t exist".to_string()));
-                    continue;
-                }
-            } else {
-                match &info {
-                    Ok(info) => {
-                        let existing_contents = fs::read_to_string(path).unwrap();
-                        if existing_contents.trim() != info.trim() {
-                            if check_codegen || !overwrite_codegen {
-                                errors.push((name, "differs from codegen".to_string()));
-                                continue;
-                            }
-                        } else {
-                            verified_count += 1;
-                            continue;
-                        }
-                    }
-                    Err(err) => {
-                        errors.push((name, err.clone()));
-                        continue;
-                    }
-                }
-            }
-
-            if let Some(dir) = path.parent() {
-                fs::create_dir_all(dir).unwrap();
-            }
-
-            let mut output_path = fs::File::create(path).unwrap();
-            match info {
-                Ok(info) => {
-                    write!(output_path, "{}", &info).unwrap();
-                    tty_println!("Saved {}", name);
-                    updated_count += 1;
-                }
-                Err(err) => {
-                    errors.push((name, err));
-                    continue;
-                }
-            }
-        }
-
-        if let Some(output_file) = output_file {
-            write_codegen_output_files(output_file, cwd, &errors);
-        }
-
-        if !errors.is_empty() {
-            println!(
-                "\nCodegen verification failed.\n\nUse hakana codegen --overwrite to regenerate\n\n{}\n\n",
-                errors
-                    .into_iter()
-                    .map(|(k, v)| format!("Error: {}\n - {}", k, v))
-                    .collect::<Vec<_>>()
-                    .join("\n")
-            );
-            exit(1);
-        }
-
-        if check_codegen {
-            tty_println!("\n{} codegen files verified!", verified_count);
-        } else {
-            tty_println!("\n{} files generated", updated_count);
-        }
-    }
-}
-
-fn do_find_paths(
-    cwd: &String,
-    all_custom_issues: FxHashSet<String>,
-    config_path: Option<&Path>,
-    sub_matches: &clap::ArgMatches,
-    analysis_hooks: Vec<Box<dyn CustomHook>>,
-    threads: u8,
-    show_progress: bool,
-    header: &str,
-    had_error: &mut bool,
-) {
-    let mut config = config::Config::new(cwd.clone(), all_custom_issues);
-    config.graph_kind = GraphKind::WholeProgram(WholeProgramKind::Query);
-
-    let config_path = config_path.unwrap();
-
-    let mut interner = Interner::default();
-
-    if config_path.exists() {
-        config
-            .update_from_file(cwd, config_path, &mut interner)
-            .ok();
-    }
-    config.allowed_issues = None;
-
-    config.security_config.max_depth =
-        if let Some(val) = sub_matches.value_of("max-depth").map(|f| f.to_string()) {
-            val.parse::<u8>().unwrap()
-        } else {
-            config.security_config.max_depth
-        };
-
-    config.hooks = analysis_hooks.into_iter().map(Arc::from).collect();
-
-    let root_dir = config.root_dir.clone();
-
-    let result = hakana_orchestrator::scan_and_analyze(
-        Vec::new(),
-        None,
-        None,
-        Arc::new(config),
-        None,
-        threads,
-        show_progress,
-        header,
-        Arc::new(interner),
-        None,
-        None,
-        None,
-        || {},
-    );
-
-    if let Ok((analysis_result, successful_run_data)) = result {
-        for (file_path, issues) in
-            analysis_result.get_all_issues(&successful_run_data.interner, &root_dir, true)
-        {
-            for issue in issues {
-                *had_error = true;
-                println!("{}", issue.format(&file_path));
-            }
-        }
-
-        if !*had_error {
-            tty_println!("\nNo security issues found!\n");
-        }
-    }
-}
-
-fn do_security_check(
-    cwd: &String,
-    all_custom_issues: FxHashSet<String>,
-    config_path: Option<&Path>,
-    sub_matches: &clap::ArgMatches,
-    analysis_hooks: Vec<Box<dyn CustomHook>>,
-    threads: u8,
-    show_progress: bool,
-    header: &str,
-    had_error: &mut bool,
-) {
-    let mut config = config::Config::new(cwd.clone(), all_custom_issues);
-    config.graph_kind = GraphKind::WholeProgram(WholeProgramKind::Taint);
-
-    let config_path = config_path.unwrap();
-
-    let mut interner = Interner::default();
-
-    if config_path.exists() {
-        config
-            .update_from_file(cwd, config_path, &mut interner)
-            .ok();
-    }
-    config.allowed_issues = None;
-
-    let output_file = sub_matches.value_of("output").map(|f| f.to_string());
-
-    config.security_config.max_depth =
-        if let Some(val) = sub_matches.value_of("max-depth").map(|f| f.to_string()) {
-            val.parse::<u8>().unwrap()
-        } else {
-            config.security_config.max_depth
-        };
-
-    config.hooks = analysis_hooks.into_iter().map(Arc::from).collect();
-
-    let root_dir = config.root_dir.clone();
-
-    let result = hakana_orchestrator::scan_and_analyze(
-        Vec::new(),
-        None,
-        None,
-        Arc::new(config),
-        None,
-        threads,
-        show_progress,
-        header,
-        Arc::new(interner),
-        None,
-        None,
-        None,
-        || {},
-    );
-
-    if let Ok((analysis_result, successful_run_data)) = result {
-        for (file_path, issues) in
-            analysis_result.get_all_issues(&successful_run_data.interner, &root_dir, true)
-        {
-            for issue in issues {
-                *had_error = true;
-                println!("{}", issue.format(&file_path));
-            }
-        }
-
-        if !*had_error {
-            tty_println!("\nNo security issues found!\n");
-        }
-
-        if let Some(output_file) = output_file {
-            write_analysis_output_files(
-                output_file,
-                None,
-                cwd,
-                &analysis_result,
-                &successful_run_data.interner,
-            );
-        }
-    }
-}
-
-async fn do_analysis(
-    sub_matches: &clap::ArgMatches,
-    all_custom_issues: FxHashSet<String>,
-    root_dir: &str,
-    mut analysis_hooks: Vec<Box<dyn CustomHook>>,
-    config_path: Option<&Path>,
-    cwd: &String,
-    cache_dir: String,
-    threads: u8,
-    show_progress: bool,
-    header: &str,
-    had_error: &mut bool,
-) {
-    use hakana_protocol::{GetIssuesRequest, Message, SocketPath};
-    use std::io::{self, Write};
-    use std::time::Duration;
-
-    let filter = sub_matches.value_of("filter").map(|f| f.to_string());
-    let standalone = sub_matches.is_present("standalone");
-    let with_server = sub_matches.is_present("with-server");
-
-    // Determine if we should use server mode
-    let project_root = Path::new(root_dir);
-    let socket_path = SocketPath::for_project(project_root);
-    let use_server = if standalone {
-        false
-    } else if with_server {
-        // --with-server: spawn server if not running
-        if !socket_path.server_exists() {
-            ServerConnection::connect_or_spawn(project_root, None)
-                .await
-                .inspect_err(|e| {
-                    println!(
-                        "Failed to spawn server: {}. Falling back to standalone analysis.",
-                        e
-                    )
-                })
-                .is_ok()
-        } else {
-            true
-        }
-    } else {
-        // Default: use server if it exists, but don't spawn one
-        socket_path.server_exists()
-    };
-
-    // Use server mode if available
-    if use_server {
-        let stdout_is_tty = io::stdout().is_terminal();
-
-        // Clear any "Waiting for server" message
-        if stdout_is_tty {
-            print!("\r\x1b[K");
-            io::stdout().flush().ok();
-        }
-        let find_unused_expressions = sub_matches.is_present("find-unused-expressions");
-        let find_unused_definitions = sub_matches.is_present("find-unused-definitions");
-
-        let request = Message::GetIssues(GetIssuesRequest {
-            filter: filter.clone(),
-            find_unused_expressions,
-            find_unused_definitions,
-            block_until_next_analysis: false,
-            send_progress_report: true,
-        });
-
-        // Poll for results, showing progress bar if analysis is in progress
-        // Reconnect for each request since server handles one request per connection
-        let pb = if stdout_is_tty {
-            let pb = ProgressBar::new(100);
-            pb.set_style(
-                ProgressStyle::with_template("{bar:40.green/yellow} {percent:>3}% - {msg}")
-                    .unwrap(),
-            );
-            pb
-        } else {
-            ProgressBar::hidden()
-        };
-
-        loop {
-            let mut client = match ClientSocket::connect(&socket_path).await {
-                Ok(c) => c,
-                Err(e) => {
-                    pb.finish_and_clear();
-                    println!("Error connecting to server: {}", e);
-                    break;
-                }
-            };
-
-            match client.request(&request).await {
-                Ok(Message::GetIssuesResult(result)) => {
-                    if result.analysis_complete {
-                        pb.finish_and_clear();
-
-                        // Print all issues
-                        for issue in result.issues {
-                            *had_error = true;
-                            println!(
-                                "{}:{}:{} - {} - {}",
-                                issue.file_path,
-                                issue.start_line,
-                                issue.start_column,
-                                issue.kind,
-                                issue.description
-                            );
-                        }
-
-                        if !*had_error {
-                            tty_println!("\nNo issues reported!\n");
-                        }
-
-                        tty_println!("\nAnalyzed {} files", result.files_analyzed);
-                        return;
-                    } else {
-                        // Update progress bar
-                        let is_analyzing = result.files_analyzed > 0;
-                        if is_analyzing {
-                            pb.set_length(result.total_files_to_analyze.max(1) as u64);
-                            pb.set_position(result.files_analyzed as u64);
-                            pb.set_message(format!(
-                                "{} ({}/{} files)",
-                                result.phase, result.files_analyzed, result.total_files_to_analyze
-                            ));
-                        } else {
-                            pb.set_length(result.total_files_to_scan.max(1) as u64);
-                            pb.set_position(result.files_scanned as u64);
-                            pb.set_message(format!(
-                                "{} ({}/{} files)",
-                                result.phase, result.files_scanned, result.total_files_to_scan
-                            ));
-                        }
-
-                        // Wait a bit before polling again (100ms for responsive UI)
-                        tokio::time::sleep(Duration::from_millis(100)).await;
-                    }
-                }
-                Ok(Message::Error(err)) => {
-                    pb.finish_and_clear();
-                    println!("Server error: {} - {}", err.code as u32, err.message);
-                    exit(1);
-                }
-                Ok(_) => {
-                    pb.finish_and_clear();
-                    println!("Unexpected response from server");
-                    exit(1);
-                }
-                Err(e) => {
-                    pb.finish_and_clear();
-                    println!("Error communicating with server: {}", e);
-                    // Fall through to standalone analysis
-                    break;
-                }
-            }
-        }
-    }
-
-    let output_file = sub_matches.value_of("output").map(|f| f.to_string());
-    let output_format = sub_matches.value_of("json-format").map(|f| f.to_string());
-
-    let ignored = sub_matches
-        .values_of("ignore")
-        .map(|values| values.map(|f| f.to_string()).collect::<FxHashSet<_>>());
-    let mut find_unused_expressions = sub_matches.is_present("find-unused-expressions");
-    let find_unused_definitions = sub_matches.is_present("find-unused-definitions");
-    let show_mixed_function_counts = sub_matches.is_present("show-mixed-function-counts");
-    let print_symbol_usages = sub_matches.is_present("print-symbol-usages");
-    let ignore_mixed_issues = sub_matches.is_present("ignore-mixed-issues");
-    let show_issue_stats = sub_matches.is_present("show-issue-stats");
-    let do_ast_diff = sub_matches.is_present("diff");
-
-    let mut issue_kinds_filter = FxHashSet::default();
-
-    let filter_issue_strings = sub_matches
-        .values_of("show-issue")
-        .map(|values| values.collect::<FxHashSet<_>>());
-
-    let show_all_issues = sub_matches.is_present("all-issues");
-
-    if let Some(filter_issue_strings) = filter_issue_strings {
-        for filter_issue_string in filter_issue_strings {
-            if let Ok(issue_kind) =
-                IssueKind::from_str_custom(filter_issue_string, &all_custom_issues)
-            {
-                if issue_kind.requires_dataflow_analysis() {
-                    find_unused_expressions = true;
-                }
-                issue_kinds_filter.insert(issue_kind);
-            } else {
-                println!("Invalid issue type {}", filter_issue_string);
-                exit(1);
-            }
-        }
-    }
-
-    let mut config = config::Config::new(root_dir.to_string(), all_custom_issues);
-    config.find_unused_expressions = find_unused_expressions;
-    config.find_unused_definitions = find_unused_definitions;
-    config.ignore_mixed_issues = ignore_mixed_issues;
-    config.ast_diff = do_ast_diff;
-
-    // filter out any hooks that cannot run in IDEs
-    analysis_hooks.retain(|h| h.run_in_ide());
-
-    config.hooks = analysis_hooks.into_iter().map(Arc::from).collect();
-
-    let config_path = config_path.unwrap();
-
-    let mut interner = Interner::default();
-
-    if config_path.exists() {
-        config
-            .update_from_file(cwd, config_path, &mut interner)
-            .ok();
-    }
-
-    // do this after we've loaded from file, as they can be overridden
-    if !issue_kinds_filter.is_empty() {
-        config.allowed_issues = Some(issue_kinds_filter);
-    } else if show_all_issues {
-        config.allowed_issues = None;
-    }
-
-    let root_dir = config.root_dir.clone();
-
-    let result = hakana_orchestrator::scan_and_analyze(
-        Vec::new(),
-        filter,
-        ignored,
-        Arc::new(config),
-        if sub_matches.is_present("no-cache") {
-            None
-        } else {
-            Some(&cache_dir)
-        },
-        threads,
-        show_progress,
-        header,
-        Arc::new(interner),
-        None,
-        None,
-        None,
-        || {},
-    );
-
-    if let Ok((analysis_result, successful_run_data)) = result {
-        if print_symbol_usages {
-            let interner = &successful_run_data.interner;
-            let codebase = &successful_run_data.codebase;
-
-            let mut symbol_definitions: BTreeMap<String, String> = BTreeMap::new();
-            let mut file_references: BTreeMap<String, Vec<String>> = BTreeMap::new();
-
-            for (symbol_id, file_paths) in &codebase.classlike_infos_defs {
-                let symbol_name = interner.lookup(symbol_id).to_string();
-                if let Some(file_path) = file_paths.first() {
-                    symbol_definitions.insert(
-                        symbol_name,
-                        file_path.get_relative_path(interner, &root_dir),
-                    );
-                }
-            }
-
-            for ((class_or_function_id, member_id), file_paths) in &codebase.functionlike_infos_defs
-            {
-                let symbol_name = if *member_id == hakana_str::StrId::EMPTY {
-                    interner.lookup(class_or_function_id).to_string()
-                } else {
-                    format!(
-                        "{}::{}",
-                        interner.lookup(class_or_function_id),
-                        interner.lookup(member_id)
-                    )
-                };
-                if let Some(file_path) = file_paths.first() {
-                    symbol_definitions.insert(
-                        symbol_name,
-                        file_path.get_relative_path(interner, &root_dir),
-                    );
-                }
-            }
-
-            for (symbol_id, file_paths) in &codebase.type_definitions_defs {
-                let symbol_name = interner.lookup(symbol_id).to_string();
-                if let Some(file_path) = file_paths.first() {
-                    symbol_definitions.insert(
-                        symbol_name,
-                        file_path.get_relative_path(interner, &root_dir),
-                    );
-                }
-            }
-
-            for (symbol_id, file_paths) in &codebase.constant_infos_defs {
-                let symbol_name = interner.lookup(symbol_id).to_string();
-                if let Some(file_path) = file_paths.first() {
-                    symbol_definitions.insert(
-                        symbol_name,
-                        file_path.get_relative_path(interner, &root_dir),
-                    );
-                }
-            }
-
-            for ((referencing_symbol, referencing_member), referenced_set) in analysis_result
-                .symbol_references
-                .symbol_references_to_symbols
-                .iter()
-                .chain(
-                    analysis_result
-                        .symbol_references
-                        .symbol_references_to_symbols_in_signature
-                        .iter(),
-                )
-            {
-                let referencing_name = if *referencing_member == hakana_str::StrId::EMPTY {
-                    interner.lookup(referencing_symbol).to_string()
-                } else {
-                    format!(
-                        "{}::{}",
-                        interner.lookup(referencing_symbol),
-                        interner.lookup(referencing_member)
-                    )
-                };
-
-                let file_path = symbol_definitions.get(&referencing_name).cloned();
-                if let Some(file_path) = file_path {
-                    let refs = file_references.entry(file_path).or_default();
-                    for (ref_symbol, ref_member) in referenced_set {
-                        let ref_name = if *ref_member == hakana_str::StrId::EMPTY {
-                            interner.lookup(ref_symbol).to_string()
-                        } else {
-                            format!(
-                                "{}::{}",
-                                interner.lookup(ref_symbol),
-                                interner.lookup(ref_member)
-                            )
-                        };
-                        if !refs.contains(&ref_name) {
-                            refs.push(ref_name);
-                        }
-                    }
-                }
-            }
-
-            for refs in file_references.values_mut() {
-                refs.sort();
-            }
-
-            let output = json!({
-                "symbol_definitions": symbol_definitions,
-                "file_references": file_references,
-            });
-
-            println!("{}", serde_json::to_string_pretty(&output).unwrap());
-            return;
-        }
-
-        for (file_path, issues) in
-            analysis_result.get_all_issues(&successful_run_data.interner, &root_dir, true)
-        {
-            for issue in issues {
-                *had_error = true;
-                println!("{}", issue.format(&file_path));
-            }
-        }
-
-        if !*had_error {
-            tty_println!("\nNo issues reported!\n");
-        }
-
-        if let Some(output_file) = output_file {
-            write_analysis_output_files(
-                output_file,
-                output_format,
-                cwd,
-                &analysis_result,
-                &successful_run_data.interner,
-            );
-        }
-
-        if show_issue_stats {
-            let mut issues_by_kind = analysis_result
-                .issue_counts
-                .into_iter()
-                .collect::<IndexMap<_, _>>();
-            issues_by_kind.sort_by(|_, a, _, b| b.cmp(a));
-
-            for (issue, count) in issues_by_kind {
-                println!("{}\t{}", issue.to_string(), count);
-            }
-        }
-
-        if show_mixed_function_counts {
-            let mut mixed_sources = analysis_result
-                .mixed_source_counts
-                .iter()
-                .map(|(k, v)| {
-                    format!(
-                        "{}\t{}",
-                        k.to_string(&successful_run_data.interner),
-                        v.len()
-                    )
-                })
-                .collect::<Vec<_>>();
-
-            mixed_sources.sort();
-
-            println!("{}", mixed_sources.join("\n"));
-        }
-    }
-}
-
-fn write_analysis_output_files(
-    output_file: String,
-    output_format: Option<String>,
-    cwd: &String,
-    analysis_result: &AnalysisResult,
-    interner: &Interner,
-) {
-    let output_path = if output_file.starts_with('/') {
-        output_file
-    } else {
-        format!("{}/{}", cwd, output_file)
-    };
-    let mut output_path = fs::File::create(Path::new(&output_path)).unwrap();
-
-    let json = match output_format {
-        Some(format) if format == "full" => {
-            let mut entries = vec![];
-
-            for (file_path, issues) in analysis_result.get_all_issues(interner, cwd, true) {
-                for issue in issues {
-                    entries.push(FullEntry::from_issue(issue, &file_path));
-                }
-            }
-
-            serde_json::to_string_pretty(&entries).unwrap()
-        }
-        Some(format) if format == "hh_client" => {
-            let mut entries = vec![];
-
-            for (file_path, issues) in analysis_result.get_all_issues(interner, cwd, true) {
-                for issue in issues {
-                    entries.push(HhClientEntry::from_issue(issue, &file_path));
-                }
-            }
-
-            serde_json::to_string_pretty(&entries).unwrap()
-        }
-        _ => {
-            let mut checkpoint_entries = vec![];
-
-            for (file_path, issues) in analysis_result.get_all_issues(interner, cwd, true) {
-                for issue in issues {
-                    checkpoint_entries.push(CheckPointEntry::from_issue(issue, &file_path));
-                }
-            }
-
-            serde_json::to_string_pretty(&checkpoint_entries).unwrap()
-        }
-    };
-    write!(output_path, "{}", json).unwrap();
-}
-
-fn write_codegen_output_files(output_file: String, cwd: &String, errors: &Vec<(String, String)>) {
-    let output_path = if output_file.starts_with('/') {
-        output_file
-    } else {
-        format!("{}/{}", cwd, output_file)
-    };
-    let mut output_path = fs::File::create(Path::new(&output_path)).unwrap();
-
-    let mut checkpoint_entries = vec![];
-
-    for (file_path, issue) in errors {
-        checkpoint_entries.push(CheckPointEntry {
-            case: "Codegen error".to_string(),
-            level: CheckPointEntryLevel::Failure,
-            filename: file_path.clone(),
-            line: 1,
-            output: issue.to_string(),
-        });
-    }
-
-    let json = serde_json::to_string_pretty(&checkpoint_entries).unwrap();
-    write!(output_path, "{}", json).unwrap();
-}
-
-fn update_files(analysis_result: &mut AnalysisResult, root_dir: &String, interner: &Interner) {
-    // Get all files that have edits
-    let files_with_edits = analysis_result.files_with_edits();
-
-    // Sort by relative path for deterministic output
-    let sorted_files: BTreeMap<String, _> = files_with_edits
-        .into_iter()
-        .map(|fp| (fp.get_relative_path(interner, root_dir), fp))
-        .collect();
-
-    for (relative_path, original_path) in sorted_files {
-        tty_println!("updating {}", relative_path);
-        let file_path = format!("{}/{}", root_dir, relative_path);
-        let file_contents = fs::read_to_string(&file_path).unwrap();
-        let mut file = File::create(&file_path).unwrap();
-
-        // Get the EditSet for this file (consumes replacements and insertions)
-        let edit_set = analysis_result.take_edits_for_file(&original_path);
-
-        // Apply edits using the unified EditSet
-        let new_contents = edit_set
-            .apply(&file_contents)
-            .unwrap_or_else(|e| panic!("Failed to apply edits to {}: {}", &file_path, e));
-
-        file.write_all(new_contents.as_bytes())
-            .unwrap_or_else(|_| panic!("Could not write file {}", &file_path));
-    }
-}
-
-/// Parse CODEOWNERS file and return patterns and exact file paths
-fn parse_codeowners_files(root_dir: &str) -> (Vec<glob::Pattern>, FxHashSet<String>) {
-    let mut codeowner_patterns = Vec::new();
-    let mut exact_files = FxHashSet::default();
-
-    let codeowners_path = format!("{}/.github/CODEOWNERS", root_dir);
-    let codeowners_content = match fs::read_to_string(&codeowners_path) {
-        Ok(content) => content,
-        Err(_) => {
-            // Try alternate location
-            let alt_path = format!("{}/CODEOWNERS", root_dir);
-            match fs::read_to_string(&alt_path) {
-                Ok(content) => content,
-                Err(_) => return (codeowner_patterns, exact_files), // No CODEOWNERS file found
-            }
-        }
-    };
-
-    for line in codeowners_content.lines() {
-        let line = line.trim();
-
-        // Skip empty lines and comments
-        if line.is_empty() || line.starts_with('#') {
-            continue;
-        }
-
-        // Parse line: "pattern @owner1 @owner2..."
-        let parts: Vec<&str> = line.split_whitespace().collect();
-        if parts.is_empty() {
-            continue;
-        }
-
-        let pattern = parts[0];
-
-        // Check if there are any owners specified
-        if parts.len() < 2 {
-            continue;
-        }
-
-        // Process pattern
-        if pattern.contains('*') || pattern.contains('?') || pattern.contains('[') {
-            // Glob pattern - compile it
-            // Convert CODEOWNERS pattern to glob pattern
-            let glob_pattern = if pattern.starts_with('/') {
-                // Absolute pattern from root
-                pattern.to_string()
-            } else if pattern.starts_with("**/") {
-                // Already in glob format
-                pattern.to_string()
-            } else {
-                // Relative pattern - match anywhere in tree
-                format!("**/{}", pattern)
-            };
-
-            if let Ok(compiled) = glob::Pattern::new(&glob_pattern) {
-                codeowner_patterns.push(compiled);
-            }
-        } else if pattern.ends_with('/') {
-            // Directory pattern with trailing slash - match all files under this directory
-            let dir_pattern = if pattern.starts_with('/') {
-                format!("{}**", pattern)
-            } else {
-                format!("**/{pattern}**")
-            };
-
-            if let Ok(compiled) = glob::Pattern::new(&dir_pattern) {
-                codeowner_patterns.push(compiled);
-            }
-        } else {
-            // Could be either an exact file or a directory without trailing slash
-            // Check if it looks like a file (has an extension like .hack or .php)
-            let is_file = pattern.ends_with(".hack")
-                || pattern.ends_with(".php")
-                || pattern.ends_with(".hhi")
-                || pattern.contains('.');
-
-            if is_file {
-                // Treat as exact file path
-                if pattern.starts_with('/') {
-                    // Pattern has leading slash - match from root
-                    exact_files.insert(pattern.to_string());
-                } else {
-                    // Pattern doesn't have leading slash - add leading slash for matching
-                    // since relative_path in matching code always has a leading slash
-                    exact_files.insert(format!("/{}", pattern));
-                }
-            } else {
-                // Treat as directory path - match all files under this directory
-                if pattern.starts_with('/') {
-                    // Absolute directory pattern from root
-                    if let Ok(compiled) = glob::Pattern::new(&format!("{}/**", pattern)) {
-                        codeowner_patterns.push(compiled);
-                    }
-                } else {
-                    // Relative directory pattern - match anywhere in tree
-                    if let Ok(compiled) = glob::Pattern::new(&format!("**/{pattern}/**")) {
-                        codeowner_patterns.push(compiled);
-                    }
-                }
-            }
-        }
-    }
-
-    (codeowner_patterns, exact_files)
-}
-
-fn do_lint(
-    sub_matches: &clap::ArgMatches,
-    root_dir: &str,
-    had_error: &mut bool,
-    custom_linters: Vec<Box<dyn hakana_lint::Linter>>,
-) {
-    use hakana_lint::{HhastLintConfig, examples};
-    use rustc_hash::FxHashMap;
-    use std::path::PathBuf;
-    use std::sync::Mutex;
-    use walkdir::WalkDir;
-
-    let config_path = sub_matches
-        .value_of("config")
-        .unwrap_or(&format!("{}/hhast-lint.json", root_dir))
-        .to_string();
-
-    let apply_fixes = sub_matches.is_present("fix");
-    let add_fixmes = sub_matches.is_present("add-fixmes");
-    let skip_codeowners = sub_matches.is_present("no-codeowners");
-    let debug_timings = sub_matches.is_present("debug-timings");
-
-    // Get specific linters to run (if provided)
-    let specific_linters = Arc::new(
-        sub_matches
-            .values_of("linter")
-            .map(|values| values.map(|s| s.to_string()).collect::<FxHashSet<_>>()),
-    );
-
-    // When --add-fixmes is used with --linter, we want to filter fixmes by linter
-    let fixme_linters = Arc::new(if add_fixmes {
-        if let Some(linters) = specific_linters.as_ref() {
-            linters.iter().cloned().collect()
-        } else {
-            Vec::new()
-        }
-    } else {
-        Vec::new()
-    });
-
-    // Parse CODEOWNERS file if skip_codeowners is enabled
-    let (codeowner_patterns, codeowner_exact_files) = if skip_codeowners {
-        parse_codeowners_files(root_dir)
-    } else {
-        (Vec::new(), FxHashSet::default())
-    };
-
-    // Load HHAST lint configuration
-    let mut hhast_config = if Path::new(&config_path).exists() {
-        match HhastLintConfig::from_file(Path::new(&config_path)) {
-            Ok(config) => config,
-            Err(e) => {
-                println!("Error loading lint config: {}", e);
-                *had_error = true;
-                return;
-            }
-        }
-    } else {
-        tty_println!(
-            "No hhast-lint.json found at {}. Using default configuration.",
-            config_path
-        );
-        HhastLintConfig::default()
-    };
-
-    // Build linter registry with all available linters
-    let mut registry = hakana_lint::linter::LinterRegistry::new();
-    registry.register(Box::new(
-        examples::no_empty_statements::NoEmptyStatementsLinter,
-    ));
-    registry.register(Box::new(
-        examples::no_whitespace_at_end_of_line::NoWhitespaceAtEndOfLineLinter,
-    ));
-    registry.register(Box::new(
-        examples::use_statement_without_kind::UseStatementWithoutKindLinter,
-    ));
-    registry.register(Box::new(
-        examples::dont_discard_new_expressions::DontDiscardNewExpressionsLinter,
-    ));
-    registry.register(Box::new(
-        examples::must_use_braces_for_control_flow::MustUseBracesForControlFlowLinter,
-    ));
-    registry.register(Box::new(examples::no_await_in_loop::NoAwaitInLoopLinter));
-    registry.register(Box::new(
-        examples::unused_use_clause::UnusedUseClauseLinter::new(),
-    ));
-
-    // Register custom linters
-    for linter in custom_linters {
-        registry.register(linter);
-    }
-
-    // Map requested linter names to HHAST names and add to config
-    if let Some(requested_linters) = specific_linters.as_ref() {
-        for linter_name in requested_linters.iter() {
-            // Try to find linter by name or HHAST name
-            let mut found = false;
-            for linter in registry.all() {
-                if let Some(hhast_name) = linter.hhast_name() {
-                    let short_name = hhast_name.split('\\').last().unwrap_or(hhast_name);
-                    if linter_name == hhast_name
-                        || linter_name == short_name
-                        || linter_name == linter.name()
-                    {
-                        if !hhast_config.extra_linters.contains(&hhast_name.to_string()) {
-                            hhast_config.extra_linters.push(hhast_name.to_string());
-                        }
-                        found = true;
-                        break;
-                    }
-                }
-            }
-            if !found {
-                eprintln!("Warning: Unknown linter '{}'", linter_name);
-            }
-        }
-    }
-    let hhast_config = Arc::new(hhast_config);
-
-    let registry = Arc::new(registry);
-
-    // Determine files to lint
-    let paths_to_lint: Vec<String> = if let Some(paths) = sub_matches.values_of("PATH") {
-        // Use explicitly provided paths
-        paths.map(|s| s.to_string()).collect()
-    } else if !hhast_config.roots.is_empty() {
-        // Use roots from config
-        hhast_config
-            .roots
-            .iter()
-            .map(|r| format!("{}/{}", root_dir, r))
-            .collect()
-    } else {
-        // Default to root_dir
-        vec![root_dir.to_string()]
-    };
-
-    // Collect all files to lint
-    let mut files_to_lint = Vec::new();
-    let mut skipped_codeowner_files = 0;
-    for base_path in paths_to_lint {
-        for entry in WalkDir::new(&base_path)
-            .follow_links(true)
-            .into_iter()
-            .filter_map(|e| e.ok())
-        {
-            let path = entry.path();
-            if !path.is_file() {
-                continue;
-            }
-
-            let path_str = path.to_string_lossy();
-            if !path_str.ends_with(".hack") && !path_str.ends_with(".php") {
-                continue;
-            }
-
-            // Skip files with codeowners if requested
-            if skip_codeowners {
-                // Canonicalize both the file path and root_dir to handle relative paths
-                let canonical_path = path.canonicalize().unwrap_or_else(|_| path.to_path_buf());
-                let canonical_root = std::path::Path::new(&root_dir)
-                    .canonicalize()
-                    .unwrap_or_else(|_| std::path::PathBuf::from(&root_dir));
-
-                let relative_path = if let Ok(rel) = canonical_path.strip_prefix(&canonical_root) {
-                    format!("/{}", rel.to_string_lossy())
-                } else {
-                    // Fallback: try to make path relative to root_dir if it's not already
-                    format!("/{}", path_str.trim_start_matches("./"))
-                };
-
-                // Check exact file match
-                let has_codeowner = codeowner_exact_files.contains(&relative_path)
-                    // Check glob patterns
-                    || codeowner_patterns.iter().any(|pattern| {
-                        // Try matching with and without leading slash
-                        pattern.matches(&relative_path)
-                            || pattern.matches(relative_path.trim_start_matches('/'))
-                    });
-
-                if has_codeowner {
-                    skipped_codeowner_files += 1;
-                    continue;
-                }
-            }
-
-            files_to_lint.push(path.to_path_buf());
-        }
-    }
-
-    if skip_codeowners && skipped_codeowner_files > 0 {
-        tty_println!(
-            "Skipped {} file(s) with codeowners",
-            skipped_codeowner_files
-        );
-    }
-
-    if files_to_lint.is_empty() {
-        tty_println!("\nNo files to lint.");
-        return;
-    }
-
-    let total_errors = Arc::new(Mutex::new(0usize));
-    let total_files = Arc::new(Mutex::new(0usize));
-    let total_fixed = Arc::new(Mutex::new(0usize));
-    let lint_output = Arc::new(Mutex::new(Vec::new()));
-    let checkpoint_entries = Arc::new(Mutex::new(Vec::new()));
-    let linter_times = Arc::new(Mutex::new(rustc_hash::FxHashMap::<
-        String,
-        std::time::Duration,
-    >::default()));
-
-    // Determine number of threads from CLI
-    let threads = if let Some(val) = sub_matches.value_of("threads").map(|f| f.to_string()) {
-        val.parse::<usize>().unwrap_or(8)
-    } else {
-        8
-    };
-
-    let mut group_size = threads;
-
-    if files_to_lint.len() < 4 * group_size {
-        group_size = 1;
-    }
-
-    let mut path_groups: FxHashMap<usize, Vec<PathBuf>> = FxHashMap::default();
-
-    for (i, path) in files_to_lint.into_iter().enumerate() {
-        let group = i % group_size;
-        path_groups.entry(group).or_insert_with(Vec::new).push(path);
-    }
-
-    let mut handles = vec![];
-
-    let root_dir = root_dir.to_string();
-
-    for (_, path_group) in path_groups {
-        let hhast_config = hhast_config.clone();
-        let registry = registry.clone();
-        let total_errors = total_errors.clone();
-        let total_files = total_files.clone();
-        let total_fixed = total_fixed.clone();
-        let lint_output = lint_output.clone();
-        let checkpoint_entries = checkpoint_entries.clone();
-        let linter_times = linter_times.clone();
-        let root_dir = root_dir.clone();
-        let fixme_linters = fixme_linters.clone();
-
-        let handle = std::thread::spawn(move || {
-            let lint_config = hakana_lint::LintConfig {
-                allow_auto_fix: apply_fixes,
-                apply_auto_fix: apply_fixes,
-                add_fixmes,
-                fixme_linters: (*fixme_linters).clone(),
-                enabled_linters: Vec::new(),
-                disabled_linters: Vec::new(),
-                root_path: Some(PathBuf::from(&root_dir)),
-            };
-
-            for path in path_group {
-                let path_str = path.to_string_lossy();
-
-                // Make path relative to root_dir for pattern matching
-                let relative_path = if let Ok(rel) = path.strip_prefix(&root_dir) {
-                    rel.to_string_lossy().to_string()
-                } else {
-                    path_str.to_string()
-                };
-
-                // Determine which linters to run for this file
-                let mut file_linters: Vec<&dyn hakana_lint::Linter> = Vec::new();
-
-                for linter in registry.all() {
-                    if let Some(hhast_name) = linter.hhast_name() {
-                        if hhast_config.is_linter_enabled(hhast_name, &relative_path) {
-                            file_linters.push(linter.as_ref());
-                        }
-                    } else {
-                        // No HHAST name, run unconditionally
-                        file_linters.push(linter.as_ref());
-                    }
-                }
-
-                if file_linters.is_empty() {
-                    continue;
-                }
-
-                // Read file contents
-                let contents = match fs::read_to_string(&path) {
-                    Ok(c) => c,
-                    Err(e) => {
-                        lint_output
-                            .lock()
-                            .unwrap()
-                            .push(format!("Error reading {}: {}", path_str, e));
-                        continue;
-                    }
-                };
-
-                *total_files.lock().unwrap() += 1;
-
-                // Create line break map for efficient offset-to-line conversions
-                let line_break_map = LineBreakMap::new(contents.as_bytes());
-
-                // Run linters
-                match hakana_lint::run_linters(&path, &contents, &file_linters, &lint_config) {
-                    Ok(result) => {
-                        // Accumulate linter times
-                        let mut times = linter_times.lock().unwrap();
-                        for (linter_name, duration) in &result.linter_times {
-                            *times
-                                .entry(linter_name.clone())
-                                .or_insert(std::time::Duration::ZERO) += *duration;
-                        }
-
-                        // Collect errors
-                        if !result.errors.is_empty() {
-                            *total_errors.lock().unwrap() += result.errors.len();
-                            for error in &result.errors {
-                                // Convert offset to line/column
-                                let (line, column) = hakana_lint::offset_to_line_column(
-                                    &line_break_map,
-                                    error.start_offset,
-                                );
-                                lint_output.lock().unwrap().push(format!(
-                                    "{}:{}:{}: {} - {}",
-                                    relative_path, line, column, error.severity, error.message
-                                ));
-
-                                // Add to checkpoint entries
-                                checkpoint_entries.lock().unwrap().push(CheckPointEntry {
-                                    case: error.linter_name.to_string(),
-                                    level: CheckPointEntryLevel::Failure,
-                                    filename: relative_path.clone(),
-                                    line: line as u32,
-                                    output: error.message.clone(),
-                                });
-                            }
-                        }
-
-                        // Apply fixes or add fixmes if requested and available
-                        if apply_fixes || add_fixmes {
-                            let mut file_ops_applied = false;
-
-                            // Apply file operations (create/delete files)
-                            if !result.file_operations.is_empty() {
-                                for file_op in &result.file_operations {
-                                    match &file_op.op_type {
-                                        hakana_lint::FileOpType::Create => {
-                                            if let Some(ref content) = file_op.content {
-                                                // Resolve relative paths relative to the source file's directory
-                                                let target_path = if file_op.path.is_absolute() {
-                                                    file_op.path.clone()
-                                                } else if let Some(parent) = path.parent() {
-                                                    parent.join(&file_op.path)
-                                                } else {
-                                                    file_op.path.clone()
-                                                };
-
-                                                // Create parent directories if they don't exist
-                                                if let Some(parent_dir) = target_path.parent() {
-                                                    if let Err(e) = fs::create_dir_all(parent_dir) {
-                                                        lint_output.lock().unwrap().push(format!(
-                                                            "Error creating directory {}: {}",
-                                                            parent_dir.display(),
-                                                            e
-                                                        ));
-                                                        continue;
-                                                    }
-                                                }
-
-                                                match fs::write(&target_path, content) {
-                                                    Ok(_) => {
-                                                        lint_output.lock().unwrap().push(format!(
-                                                            "Created: {}",
-                                                            target_path.display()
-                                                        ));
-                                                        file_ops_applied = true;
-                                                    }
-                                                    Err(e) => {
-                                                        lint_output.lock().unwrap().push(format!(
-                                                            "Error creating {}: {}",
-                                                            target_path.display(),
-                                                            e
-                                                        ));
-                                                    }
-                                                }
-                                            }
-                                        }
-                                        hakana_lint::FileOpType::Delete => {
-                                            let target_path = if file_op.path.is_absolute() {
-                                                file_op.path.clone()
-                                            } else if let Some(parent) = path.parent() {
-                                                parent.join(&file_op.path)
-                                            } else {
-                                                file_op.path.clone()
-                                            };
-
-                                            match fs::remove_file(&target_path) {
-                                                Ok(_) => {
-                                                    lint_output.lock().unwrap().push(format!(
-                                                        "Deleted: {}",
-                                                        target_path.display()
-                                                    ));
-                                                    file_ops_applied = true;
-                                                }
-                                                Err(e) => {
-                                                    lint_output.lock().unwrap().push(format!(
-                                                        "Error deleting {}: {}",
-                                                        target_path.display(),
-                                                        e
-                                                    ));
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-
-                            // Apply text edits to the source file
-                            if let Some(fixed_source) = result.modified_source {
-                                match fs::write(&path, fixed_source) {
-                                    Ok(_) => {
-                                        *total_fixed.lock().unwrap() += 1;
-                                        if add_fixmes {
-                                            lint_output
-                                                .lock()
-                                                .unwrap()
-                                                .push(format!("Added fixmes: {}", relative_path));
-                                        } else {
-                                            lint_output
-                                                .lock()
-                                                .unwrap()
-                                                .push(format!("Fixed: {}", relative_path));
-                                        }
-                                    }
-                                    Err(e) => {
-                                        lint_output.lock().unwrap().push(format!(
-                                            "Error writing {} to {}: {}",
-                                            if add_fixmes { "fixmes" } else { "fixes" },
-                                            path_str,
-                                            e
-                                        ));
-                                    }
-                                }
-                            } else if file_ops_applied {
-                                // If we only did file operations, count that as a fix
-                                *total_fixed.lock().unwrap() += 1;
-                            }
-                        }
-                    }
-                    Err(e) => {
-                        lint_output
-                            .lock()
-                            .unwrap()
-                            .push(format!("Error linting {}: {}", path_str, e));
-                    }
-                }
-            }
-        });
-
-        handles.push(handle);
-    }
-
-    // Wait for all threads to complete
-    for handle in handles {
-        handle.join().unwrap();
-    }
-
-    // Print all output (sorted alphabetically)
-    let mut output = lint_output.lock().unwrap().clone();
-    output.sort();
-    for line in output.iter() {
-        println!("{}", line);
-    }
-
-    let final_errors = *total_errors.lock().unwrap();
-    let final_files = *total_files.lock().unwrap();
-    let final_fixed = *total_fixed.lock().unwrap();
-
-    if final_errors > 0 {
-        *had_error = true;
-    }
-
-    // Print summary
-    if final_errors > 0 {
-        tty_println!(
-            "\nFound {} lint issue(s) in {} file(s)",
-            final_errors,
-            final_files
-        );
-        if (apply_fixes || add_fixmes) && final_fixed > 0 {
-            if add_fixmes {
-                tty_println!("Added fixmes to {} file(s)", final_fixed);
-            } else {
-                tty_println!("Fixed {} file(s)", final_fixed);
-            }
-        }
-    } else {
-        tty_println!("\nNo lint issues found! Checked {} file(s).", final_files);
-    }
-
-    // Print linter timing statistics (only if --debug-timings is passed)
-    if debug_timings {
-        let times = linter_times.lock().unwrap();
-        if !times.is_empty() {
-            println!("\nLinter timing statistics:");
-            let mut sorted_times: Vec<_> = times.iter().collect();
-            sorted_times.sort_by(|a, b| b.1.cmp(a.1)); // Sort by duration, descending
-
-            let total_time: std::time::Duration = times.values().sum();
-
-            for (linter_name, duration) in sorted_times {
-                let secs = duration.as_secs_f64();
-                let percentage = if total_time.as_secs_f64() > 0.0 {
-                    (secs / total_time.as_secs_f64()) * 100.0
-                } else {
-                    0.0
-                };
-                println!(
-                    "  {:<50} {:>8.3}s ({:>5.1}%)",
-                    linter_name, secs, percentage
-                );
-            }
-            println!("  {:<50} {:>8.3}s", "Total", total_time.as_secs_f64());
-        }
-    }
-
-    // Write checkpoint results if output file is specified
-    if let Some(output_path) = sub_matches.value_of("output") {
-        let entries = checkpoint_entries.lock().unwrap();
-        let json = serde_json::to_string_pretty(&*entries).unwrap();
-        let mut output_file = fs::File::create(Path::new(output_path)).unwrap();
-        write!(output_file, "{}", json).unwrap();
-    }
-}
-
-async fn do_server(
-    sub_matches: &clap::ArgMatches,
-    root_dir: &str,
-    threads: u8,
-    header: &str,
-    analysis_hooks: Vec<Box<dyn CustomHook>>,
-) {
-    use hakana_protocol::{Message, ShutdownRequest, SocketPath, StatusRequest};
-    use hakana_server::{Server, ServerConfig};
-
-    let socket_path = SocketPath::for_project(Path::new(root_dir));
-
-    // Handle --status flag
-    if sub_matches.is_present("status") {
-        if !socket_path.server_exists() {
-            println!("Server not running");
-            return;
-        }
-
-        match ClientSocket::connect(&socket_path).await {
-            Ok(mut client) => {
-                let response = client.request(&Message::Status(StatusRequest)).await;
-                match response {
-                    Ok(Message::StatusResult(status)) => {
-                        println!("Server Status:");
-                        println!("  Ready: {}", status.ready);
-                        println!("  Files: {}", status.files_count);
-                        println!("  Symbols: {}", status.symbols_count);
-                        println!("  Uptime: {}s", status.uptime_secs);
-                        println!("  Analysis in progress: {}", status.analysis_in_progress);
-                        println!("  Project root: {}", status.project_root);
-                    }
-                    Ok(_) => println!("Unexpected response from server"),
-                    Err(e) => println!("Error communicating with server: {}", e),
-                }
-            }
-            Err(e) => {
-                println!("Cannot connect to server: {}", e);
-            }
-        }
-        return;
-    }
-
-    // Handle --stop flag
-    if sub_matches.is_present("stop") {
-        if !socket_path.server_exists() {
-            println!("Server not running");
-            return;
-        }
-
-        match ClientSocket::connect(&socket_path).await {
-            Ok(mut client) => match client.send(&Message::Shutdown(ShutdownRequest)).await {
-                Ok(_) => println!("Shutdown signal sent"),
-                Err(e) => println!("Error sending shutdown: {}", e),
-            },
-            Err(e) => {
-                println!("Cannot connect to server: {}", e);
-            }
-        }
-        return;
-    }
-
-    // Start server
-    let config_path = sub_matches
-        .value_of("config")
-        .map(|s| s.to_string())
-        .unwrap_or_else(|| format!("{}/hakana.json", root_dir));
-
-    // Convert hooks to Arc for reuse across multiple analyses
-    let plugins: Vec<Arc<dyn CustomHook>> = analysis_hooks.into_iter().map(Arc::from).collect();
-
-    let server_config = ServerConfig {
-        root_dir: root_dir.to_string(),
-        threads,
-        config_path: Some(config_path),
-        plugins,
-        header: header.to_string(),
-        chaos_monkey: None,
-    };
-
-    match Server::new(server_config) {
-        Ok(mut server) => {
-            tty_println!("Starting hakana server...");
-            tty_println!("Socket: {}", server.socket_path().path().display());
-            if let Err(e) = server.run().await {
-                println!("Server error: {}", e);
-                exit(1);
-            }
-        }
-        Err(e) => {
-            println!("Failed to start server: {}", e);
-            exit(1);
-        }
-    }
-}
-
-/// Cyclomatic complexity analysis
-fn do_cyclomatic_complexity(
-    sub_matches: &clap::ArgMatches,
-    all_custom_issues: FxHashSet<String>,
-    root_dir: &str,
-    config_path: Option<&Path>,
-    cwd: &String,
-    threads: u8,
-    show_progress: bool,
-    header: &str,
-    had_error: &mut bool,
-) {
-    let mut config = config::Config::new(root_dir.to_string(), all_custom_issues);
-
-    let config_path = config_path.unwrap();
-
-    let mut interner = Interner::default();
-
-    if config_path.exists() {
-        config
-            .update_from_file(cwd, config_path, &mut interner)
-            .ok();
-    }
-    config.allowed_issues = None;
-
-    config.analyze_cyclomatic_complexity = true;
-
-    config.cyclomatic_complexity_file_patterns = sub_matches
-        .values_of("filter")
-        .map(|values| {
-            values
-                .map(|f| {
-                    glob::Pattern::new(&format!("{}/{}", cwd, f)).expect("invalid filter pattern")
-                })
-                .collect()
-        })
-        .unwrap_or_default();
-
-    let threshold = sub_matches
-        .value_of("threshold")
-        .map(|t| t.parse::<u32>().unwrap_or(10))
-        .unwrap_or(10);
-    config.cyclomatic_complexity_threshold = threshold;
-
-    let output_file = sub_matches.value_of("output").map(|f| f.to_string());
-
-    let result = hakana_orchestrator::scan_and_analyze(
-        Vec::new(),
-        None,
-        None,
-        Arc::new(config),
-        None,
-        threads,
-        show_progress,
-        header,
-        Arc::new(interner),
-        None,
-        None,
-        None,
-        || {},
-    );
-
-    if let Ok((analysis_result, successful_run_data)) = result {
-        let mut results = analysis_result.cyclomatic_complexity;
-        results.sort_by(|a, b| b.cmp(&successful_run_data.interner, &a));
-
-        let total_functions = successful_run_data.codebase.functionlike_infos.len();
-
-        if results.is_empty() {
-            tty_println!("No functions exceed complexity threshold of {}", threshold);
-        } else {
-            tty_println!(
-                "{}/{} functions exceed the complexity threshold:\n",
-                results.len(),
-                total_functions
-            );
-
-            for function_complexity in &results {
-                println!(
-                    "{}",
-                    function_complexity.to_string(&successful_run_data.interner)
-                );
-            }
-        }
-
-        if let Some(output_path_str) = output_file {
-            let report = json!({
-                "threshold": threshold,
-                "total_functions": total_functions,
-                "functions_over_threshold": results.len(),
-                "results": results.iter().map(|c| c.to_json(&successful_run_data.interner)).collect::<Vec<_>>(),
-            });
-
-            let json = serde_json::to_string_pretty(&report).unwrap();
-            let mut out = fs::File::create(Path::new(&output_path_str)).unwrap();
-            write!(out, "{}", json).unwrap();
-            tty_println!("\nResults written to: {}", output_path_str);
-        }
-
-        if !results.is_empty() {
-            *had_error = true;
-        }
     }
 }
